requirements.txt
streamlit>=1.35.0
tensorflow>=2.14.0
numpy
pillow
matplotlib
scikit-learn